+++ src/gpre/cob.cpp
@@ -2508,7 +2508,7 @@ static void gen_event_wait( const act* action)
 	PAT args;
 	args.pat_database = database;
 	args.pat_vector1 = status_vector(action);
-	args.pat_value1 = (int) ident;
+	args.pat_value1 = ident;
 	args.pat_string2 = ISC_EVENT_WAIT;
 	args.pat_string3 = ISC_EVENT_COUNTS;
 	args.pat_string4 = names[isc_a_pos];
